+++ packages/jsenv-plugin-bundling/src/js_module/bundle_js_modules.js
@@ -338,7 +338,7 @@ const rollupPluginJsenv = ({
         // happen when excluded by referenceAnalysis.include
         return { id: url, external: true };
       }
-      if (urlInfo.mustIgnore) {
+      if (urlInfo.url.startsWith("ignore:")) {
         return { id: url, external: true };
       }
       const filePath = fileUrlConverter.asFilePath(url);