+++ packages/test/src/coverage/coverage_reporter_text_log.js
@@ -1,12 +1,12 @@
-import { requireFromJsenv } from "@jsenv/core/src/helpers/require_from_jsenv.js";
+import { importWithRequire } from "../helpers/import_with_require.js";
 import { istanbulCoverageMapFromCoverage } from "./istanbul_coverage_map_from_coverage.js";
 
 export const generateCoverageTextLog = (
   coverage,
   { coverageReportSkipEmpty, coverageReportSkipFull },
 ) => {
-  const libReport = requireFromJsenv("istanbul-lib-report");
-  const reports = requireFromJsenv("istanbul-reports");
+  const libReport = importWithRequire("istanbul-lib-report");
+  const reports = importWithRequire("istanbul-reports");
 
   const context = libReport.createContext({
     coverageMap: istanbulCoverageMapFromCoverage(coverage),