+++ src/plugins/protocol_http/jsenv_plugin_protocol_http.js
@@ -3,14 +3,15 @@ export const jsenvPluginProtocolHttp = () => {
     name: "jsenv:protocol_http",
     appliesDuring: "*",
     redirectReference: (reference) => {
+      // TODO: according to some pattern matching jsenv could be allowed
+      // to fetch and transform http urls
       if (
         reference.url.startsWith("http:") ||
         reference.url.startsWith("https:")
       ) {
-        reference.mustIgnore = true;
+        return `ignore:${reference.url}`;
       }
-      // TODO: according to some pattern matching jsenv could be allowed
-      // to fetch and transform http urls
+      return null;
     },
   };
 };