+++ src/plugins/reference_analysis/jsenv_plugin_reference_analysis.js
@@ -1,5 +1,3 @@
-import { URL_META } from "@jsenv/url-meta";
-
 import { jsenvPluginReferenceExpectedTypes } from "./jsenv_plugin_reference_expected_types.js";
 import { jsenvPluginDirectoryReferenceAnalysis } from "./directory/jsenv_plugin_directory_reference_analysis.js";
 import { jsenvPluginDataUrlsAnalysis } from "./data_urls/jsenv_plugin_data_urls_analysis.js";
@@ -9,21 +7,12 @@ import { jsenvPluginCssReferenceAnalysis } from "./css/jsenv_plugin_css_referenc
 import { jsenvPluginJsReferenceAnalysis } from "./js/jsenv_plugin_js_reference_analysis.js";
 
 export const jsenvPluginReferenceAnalysis = ({
-  include,
-  supportedProtocols = ["file:", "data:", "virtual:", "http:", "https:"],
-
-  ignoreProtocol = "remove",
   inlineContent = true,
   inlineConvertedScript = false,
   fetchInlineUrls = true,
   allowEscapeForVersioning = false,
 }) => {
   return [
-    jsenvPluginReferenceAnalysisInclude({
-      include,
-      supportedProtocols,
-      ignoreProtocol,
-    }),
     jsenvPluginDirectoryReferenceAnalysis(),
     jsenvPluginHtmlReferenceAnalysis({
       inlineContent,
@@ -41,83 +30,6 @@ export const jsenvPluginReferenceAnalysis = ({
       : []),
     jsenvPluginReferenceExpectedTypes(),
   ];
-};
-
-const jsenvPluginReferenceAnalysisInclude = ({
-  include,
-  supportedProtocols,
-  ignoreProtocol,
-}) => {
-  // eslint-disable-next-line no-unused-vars
-  let getIncludeInfo = (url) => undefined;
-
-  return {
-    name: "jsenv:reference_analysis_include",
-    appliesDuring: "*",
-    init: ({ rootDirectoryUrl }) => {
-      if (include) {
-        const associations = URL_META.resolveAssociations(
-          { include },
-          rootDirectoryUrl,
-        );
-        getIncludeInfo = (url) => {
-          const { include } = URL_META.applyAssociations({
-            url,
-            associations,
-          });
-          return include;
-        };
-      }
-    },
-    redirectReference: (reference) => {
-      if (reference.mustIgnore !== undefined) {
-        return;
-      }
-      if (
-        reference.specifier[0] === "#" &&
-        // For Html, css and in general "#" refer to a resource in the page
-        // so that urls must be kept intact
-        // However for js import specifiers they have a different meaning and we want
-        // to resolve them (https://nodejs.org/api/packages.html#imports for instance)
-        reference.type !== "js_import"
-      ) {
-        reference.mustIgnore = true;
-        return;
-      }
-      if (reference.url.startsWith("ignore:")) {
-        reference.mustIgnore = true;
-        if (ignoreProtocol === "remove") {
-          reference.specifier = reference.specifier.slice("ignore:".length);
-        }
-        return;
-      }
-      const includeInfo = getIncludeInfo(reference.url);
-      if (includeInfo === true) {
-        reference.mustIgnore = false;
-        return;
-      }
-      if (includeInfo === false) {
-        reference.mustIgnore = true;
-        return;
-      }
-      const { protocol } = new URL(reference.url);
-      const protocolIsSupported = supportedProtocols.some(
-        (supportedProtocol) => protocol === supportedProtocol,
-      );
-      if (!protocolIsSupported) {
-        reference.mustIgnore = true;
-      }
-    },
-    formatReference: (reference) => {
-      if (
-        ignoreProtocol === "inject" &&
-        reference.mustIgnore &&
-        !reference.url.startsWith("ignore:")
-      ) {
-        reference.specifier = `ignore:${reference.specifier}`;
-      }
-    },
-  };
 };
 
 const jsenvPluginInlineContentFetcher = () => {