+++ packages/jsenv-plugin-toolbar/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.1.0
+
+Add a button to show/hide dev server ribbon
+
 # 1.0.0
 
 First version